langchain/openai-translator/requirements.txt
pdfplumber
simplejson
requests
PyYAML
pillow
reportlab
pandas
loguru
openai
# langchain
# gradio
flask
# langchain-community
langchain==0.1.16
langchain-openai==0.1.1
langchain-experimental==0.0.55
gradio==4.22.0
httpx
httpx-sse
PyJWT
zhipuai==2.0.1.20240427
